add class reg

+++ simulator/read_file.cpp
@@ -19,12 +19,12 @@ void read_file(std::vector<uint>& iimage, uint& pc, uint& amt_inst){
         pc=0;
         for(int i=0;i<4;++i){
             pc<<=8;
-            pc+=iimage_char[i];
+            pc+=(uint)((uchar)(iimage_char[i]));
         }
         amt_inst=0;
         for(int i=4;i<8;++i){
             amt_inst<<=8;
-            amt_inst+=iimage_char[i];
+            amt_inst+=(uint)((uchar)(iimage_char[i]));
         }
         uint temp;
         iimage.clear();
@@ -32,7 +32,7 @@ void read_file(std::vector<uint>& iimage, uint& pc, uint& amt_inst){
             temp=0;
             for(int j=i;j<i+4;++j){
                 temp<<=8;
-                temp+=iimage_char[j];
+                temp+=(uint)((uchar)(iimage_char[j]));
             }
             iimage.push_back(temp);
         }
